refactor: move channels_v2/registry.py to channels/registry.py
